chore: re-use extract-user methods
