+++ src/lib/features/project/createProjectService.ts
@@ -1,5 +1,5 @@
 import { Db, IUnleashConfig } from '../../server-impl';
-import EventStore from '../../db/event-store';
+import EventStore from '../events/event-store';
 import GroupStore from '../../db/group-store';
 import { AccountStore } from '../../db/account-store';
 import EnvironmentStore from '../project-environments/environment-store';
