+++ src/lib/routes/admin-api/event.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 import { IUnleashConfig } from '../../types/option';
 import { IUnleashServices } from '../../types/services';
-import EventService from '../../services/event-service';
+import EventService from '../../features/events/event-service';
 import { ADMIN, NONE } from '../../types/permissions';
 import { IEvent, IEventList } from '../../types/events';
 import Controller from '../controller';
